refactor: require admitted created-target execution
