+++ .changeset/safe-mcp-action-preflight.md
@@ -27,7 +27,8 @@ Move generic MCP action targets, idempotency fingerprints, and approval prefligh
 behind a discoverable server-owned Tool contract. Migrated packages resolve ledger
 targets from validated input, approval-required calls return structured server-issued
 approval metadata, and exact retries are validated against the stored command and
-principal. Unmigrated execute actions temporarily retain their previous advertised
-invocation contract for a merge-safe package rollout. Validate graph risk against
-the loaded Tool tier before release and keep the Operator MCP health check from
-accepting startup failures.
+principal. Route every handler-created target through the selected Tool admission,
+remove the raw created-target executor from the public package surface, and reject
+future package-level bypasses during manifest convergence. Validate graph risk
+against the loaded Tool tier before release and keep the Operator MCP health check
+from accepting startup failures.
