+++ packages/action-ledger/README.md
@@ -36,43 +36,25 @@ canonical target does not exist before dispatch, so the caller cannot supply it 
 preflight cannot share the domain transaction. A created-target handler must implement the
 `handler-command-claim-v1` contract: claim a stable pre-create command identity and fingerprint
 before mutation, reject same-key/different-command reuse, replay a typed immutable result
-reference, and atomically append the canonical generated-target result. Approval-bearing created
-commands use MCP's request-scoped `handlerActionPolicy` context and are validated inside that same
-transaction before the claim or domain mutation.
+reference, and atomically append the canonical generated-target result.
 
-Package handlers implement that contract with `executeCreatedTargetCommand` from
-`@voyant-travel/action-ledger/created-command`. It owns the transaction and keeps the claim
-opaque; domain callbacks receive only its exact transaction handle:
+Package handlers implement that contract with `executeAdmittedCreatedTargetCommand` from
+`@voyant-travel/action-ledger/created-command`. The executor derives immutable action identity,
+route, scope, and fingerprint from the selected graph action and admitted Tool invocation. Package
+code supplies only domain target metadata and mutation callbacks:
 
 ```ts
-const command = {
-  actionName: "relationship.person.create",
-  actionVersion: "v1",
-  commandTarget: {
-    type: "relationship-person-create-command",
-    id: commandId,
-  },
-  canonicalTargetType: "relationship-person",
-  resultReferenceType: "relationship-person-ref",
-  commandInput: input,
-  capabilityId: "relationships:person:create",
-  capabilityVersion: "v1",
-  evaluatedRisk: "high",
-  approvalPolicy: "none",
-  approvalReasonCode: null,
-}
-const fingerprint = await buildCreatedTargetCommandFingerprint(command)
-
-return executeCreatedTargetCommand(
-  db,
+return executeAdmittedCreatedTargetCommand(
   {
+    db,
     context,
-    ...command,
-    idempotency: {
-      scope: `relationships.create_person:${principalId}`,
-      key: idempotencyKey,
-      fingerprint,
-    },
+    admitted: ctx.handlerActionPolicy,
+    idempotencyKey: input.idempotencyKey,
+    commandTargetType: "relationship-person-create-command",
+    canonicalTargetType: "relationship-person",
+    resultReferenceType: "relationship-person-ref",
+    commandInput: input,
+    evaluatedRisk: "high",
   },
   {
     async create(tx) {
@@ -86,7 +68,10 @@ return executeCreatedTargetCommand(
 )
 ```
 
-The executor requires a transaction-capable database, holds a Postgres transaction-scoped
+The executor fails closed unless the admitted action is handler-enforced, requires the ledger,
+declares the exact created-target protocol, matches the supplied target/risk metadata, carries an
+admitted idempotency key, and has `approval: "never"`. It requires a transaction-capable database,
+holds a Postgres transaction-scoped
 advisory lock for the idempotency scope and key, appends the requested command identity before
 calling domain code, re-reads that opaque claim, and appends the canonical result before commit.
 Exact replay validates full principal, tenant, workflow, capability, authorization, and approval
@@ -95,30 +80,12 @@ committed claim without a result throws `ActionLedgerCreatedCommandReplayIncompl
 malformed or inconsistent result metadata throws
 `ActionLedgerCreatedCommandReplayCorruptError`. Neither condition is dispatched again.
 
-`buildCreatedTargetCommandFingerprint` covers the command identity and input,
-`canonicalTargetType`, `resultReferenceType`, and typed risk/capability/approval/reason metadata.
-The executor derives those fields from the same top-level input and fails before the transaction
-if the supplied digest drifts. Principal admission uses `mapActionLedgerRequestContext`; mismatched
-caller types cannot smuggle an agent or API-token identity into the ledger.
-
-For a selected action with `approval: "required"`, `request_action_approval` fingerprints the
-declared command target type, canonical target type, result-reference type, capability/version,
-risk, reason, exact Tool capability, and command input with
-`buildCreatedTargetCommandFingerprint`. A multi-Tool action must name its selected
-`toolCapabilityId`; a single-Tool action derives it unambiguously. The package handler passes the
-fresh `ctx.handlerActionPolicy.invocation` approval id, idempotency key, fingerprint, and reason
-into `approvalControls`, uses `ctx.handlerActionPolicy.capabilityId` as `routeOrToolName`, and
-supplies the selected policy name.
-
-Inside its owned transaction, `executeCreatedTargetCommand` locks the approval before the command
-scope. First execution validates the approved request's fingerprint, command target,
-capability/version, risk, policy snapshot, reason, idempotency key, requester, assignee/decider,
-and expiry, then rejects any prior claim linked to the same approved request even if its command
-scope differs. It derives claim causation and approval linkage from the validated rows. An exact
-linked replay checks immutable approval/request/claim/result continuity without re-running expiry
-authorization, so a completed command remains replayable after its approval expires. Direct
-caller-supplied `approvalId` or `causationActionId` fails closed. Conditional policy remains
-unsupported until a domain evaluator contract exists.
+The internal command fingerprint covers action identity and input, canonical and command target
+types, result-reference type, and typed risk/capability/approval metadata. Principal admission uses
+`mapActionLedgerRequestContext`; mismatched caller types cannot smuggle an agent or API-token
+identity into the ledger. Created-target actions requiring approval remain unavailable until an
+admitted approval-bearing executor exists; package code cannot opt into a lower-level execution
+surface.
 
 Existing-target actions declare `commandTargetField` when their Tool input already carries the
 domain target id. The registry resolves that field from already parsed input and the generic gate
