+++ packages/commerce/src/created-target-command.ts
@@ -1,26 +1,21 @@
 import {
   type ActionLedgerRequestContextValues,
-  buildCreatedTargetIdempotencyScope,
+  type ExecuteAdmittedCreatedTargetCommandInput,
   type ExecuteCreatedTargetCommandHandlers,
-  type ExecuteCreatedTargetCommandInput,
   type ExecuteCreatedTargetCommandResult,
-  executeCreatedTargetCommand,
+  executeAdmittedCreatedTargetCommand,
   mapActionLedgerRequestContext,
 } from "@voyant-travel/action-ledger"
 import { ToolError, type ToolHandlerActionPolicyContext } from "@voyant-travel/tools"
 import type { PostgresJsDatabase } from "drizzle-orm/postgres-js"
 
-import {
-  buildCommerceCreatedTargetFingerprint,
-  type COMMERCE_CREATED_TARGET_POLICIES,
-} from "./created-target-policy.js"
+import type { COMMERCE_CREATED_TARGET_POLICIES } from "./created-target-policy.js"
 
 type CommerceCreatedTargetPolicy =
   (typeof COMMERCE_CREATED_TARGET_POLICIES)[keyof typeof COMMERCE_CREATED_TARGET_POLICIES]
 
 type CommerceCreatedCommandExecutor = (
-  db: PostgresJsDatabase,
-  input: ExecuteCreatedTargetCommandInput & { resultReferenceType: string },
+  input: ExecuteAdmittedCreatedTargetCommandInput<string>,
   handlers: ExecuteCreatedTargetCommandHandlers<{ id: string }, string>,
 ) => Promise<ExecuteCreatedTargetCommandResult<{ id: string }, string>>
 
@@ -32,52 +27,24 @@ export async function executeCommerceCreate(
   commandInput: unknown,
   admitted: ToolHandlerActionPolicyContext,
   create: (tx: PostgresJsDatabase) => Promise<{ id: string }>,
-  executor: CommerceCreatedCommandExecutor = executeCreatedTargetCommand,
+  executor: CommerceCreatedCommandExecutor = executeAdmittedCreatedTargetCommand,
 ) {
   const principal = mapActionLedgerRequestContext(context)
   if (principal.principalId === "unknown_request") {
     throw new TypeError("Commerce created-target commands require a concrete principal")
   }
-  const idempotencyKey = admittedCreatedCommandIdempotencyKey(admitted, legacyIdempotencyKey)
-  const selectedActionName = admitted.actionPolicy.capabilityId
-  const selectedActionVersion = admitted.actionPolicy.version
-  const fingerprint = await buildCommerceCreatedTargetFingerprint(
-    {
-      ...policy,
-      actionName: selectedActionName,
-      actionVersion: selectedActionVersion,
-      capabilityId: selectedActionName,
-      capabilityVersion: selectedActionVersion,
-    } as CommerceCreatedTargetPolicy,
-    idempotencyKey,
-    commandInput,
-  )
-  const scope = await buildCreatedTargetIdempotencyScope({
-    actionName: selectedActionName,
-    actionVersion: selectedActionVersion,
-    principalType: principal.principalType,
-    principalId: principal.principalId,
-    organizationId: principal.organizationId,
-  })
+  admittedCreatedCommandIdempotencyKey(admitted, legacyIdempotencyKey)
   return executor(
-    db,
     {
+      db,
       context,
-      actionName: selectedActionName,
-      actionVersion: selectedActionVersion,
-      actionKind: "create",
-      evaluatedRisk: policy.evaluatedRisk,
-      commandTarget: { type: policy.commandTargetType, id: idempotencyKey },
+      admitted,
+      idempotencyKey: legacyIdempotencyKey,
+      commandTargetType: policy.commandTargetType,
       canonicalTargetType: policy.canonicalTargetType,
       resultReferenceType: policy.resultReferenceType,
-      capabilityId: selectedActionName,
-      capabilityVersion: selectedActionVersion,
-      approvalPolicy: policy.approvalPolicy,
-      approvalReasonCode: policy.approvalReasonCode,
       commandInput,
-      routeOrToolName: admitted.capabilityId,
-      authorizationSource: "selected_graph_mcp_handler",
-      idempotency: { scope, key: idempotencyKey, fingerprint },
+      evaluatedRisk: policy.evaluatedRisk,
     },
     {
       async create(tx) {
