+++ packages/cruises/tests/unit/created-target-policy.test.ts
@@ -64,16 +64,15 @@ describe("cruise ship created-target command", () => {
     const routes: Array<string | null | undefined> = []
     const actionIdentities: Array<[string, string, string, string]> = []
     const executor: NonNullable<Parameters<typeof executeCruiseShipCreate>[6]> = async (
-      _db,
       input,
       handlers,
     ) => {
-      routes.push(input.routeOrToolName)
+      routes.push(input.admitted.capabilityId)
       actionIdentities.push([
-        input.actionName,
-        input.actionVersion,
-        input.capabilityId,
-        input.capabilityVersion,
+        input.admitted.actionPolicy.capabilityId,
+        input.admitted.actionPolicy.version,
+        input.admitted.actionPolicy.capabilityId,
+        input.admitted.actionPolicy.version,
       ])
       if (!createdId) {
         const mutation = await handlers.create(tx)
